Make StoreEntry part of public API via hydra_zen.typing
